Consolidate app e2e smoke data
