--- test/e2e/app/src/config-smoke.ts
@@ -1,24 +1,0 @@
-import type { LinkSmokeCase } from "../helpers";
-
-export const configLinks: LinkSmokeCase[] = [
-  { href: "/config/integrations", label: "Devices & services" },
-  { href: "/config/automation", label: "Automations & scenes" },
-  { href: "/config/areas", label: "Areas, labels & zones" },
-  { href: "/config/apps", label: "Apps" },
-  { href: "/config/lovelace/dashboards", label: "Dashboards" },
-  { href: "/config/voice-assistants", label: "Voice assistants" },
-  { href: "/config/matter", label: "Matter" },
-  { href: "/config/zha", label: "Zigbee" },
-  { href: "/config/zwave_js", label: "Z-Wave" },
-  { href: "/knx", label: "KNX" },
-  { href: "/config/thread", label: "Thread" },
-  { href: "/config/bluetooth", label: "Bluetooth" },
-  { href: "/config/infrared", label: "Infrared" },
-  { href: "/config/radio-frequency", label: "Radio frequency" },
-  { href: "/insteon", label: "Insteon" },
-  { href: "/config/tags", label: "Tags" },
-  { href: "/config/person", label: "People" },
-  { href: "/config/system", label: "System" },
-  { href: "/config/tools", label: "Tools" },
-  { href: "/config/info", label: "About" },
-];
